+++ src/parameter_number_message_scanner.rs
@@ -5,10 +5,13 @@ use crate::{
 
 /// Scanner for detecting (N)RPN messages in a stream of short messages without polling.
 ///
-/// Supports the following value message sequences:
+/// Supports the following message sequences (`x` and `y` represent the bytes that make up the
+/// parameter number):
 ///
-/// - `[MSB]`: Interpreted as 7-bit message.
-/// - `[LSB, MSB]`: Interpreted as 14-bit message.
+/// - `[x, y, MSB]`: Interpreted as 7-bit message.
+/// - `[x, y, LSB, MSB]`: Interpreted as 14-bit message.
+/// - `[x, y, MSB, MSB, ...]`: Interpreted as 7-bit messages.
+/// - `[x, y, LSB, MSB, LSB, MSB, ...]`: Interpreted as 14-bit messages.
 ///
 /// # Example
 ///
@@ -71,7 +74,7 @@ struct ScannerForOneChannel {
 }
 
 impl ScannerForOneChannel {
-    fn feed(&mut self, msg: &impl ShortMessage) -> Option<ParameterNumberMessage> {
+    pub fn feed(&mut self, msg: &impl ShortMessage) -> Option<ParameterNumberMessage> {
         match msg.to_structured() {
             StructuredShortMessage::ControlChange {
                 channel,
@@ -90,7 +93,7 @@ impl ScannerForOneChannel {
         }
     }
 
-    fn reset(&mut self) {
+    pub fn reset(&mut self) {
         self.number_msb = None;
         self.number_lsb = None;
         self.is_registered = false;
@@ -132,23 +135,15 @@ impl ScannerForOneChannel {
         let number_lsb = self.number_lsb?;
         let number_msb = self.number_msb?;
         let number = build_14_bit_value_from_two_7_bit_values(number_msb, number_lsb);
-        let msg = if self.is_registered {
-            match self.value_lsb {
-                Some(value_lsb) => ParameterNumberMessage::registered_14_bit(
-                    channel,
-                    number,
-                    build_14_bit_value_from_two_7_bit_values(value_msb, value_lsb),
-                ),
-                None => ParameterNumberMessage::registered_7_bit(channel, number, value_msb),
-            }
-        } else {
-            match self.value_lsb {
-                Some(value_lsb) => ParameterNumberMessage::non_registered_14_bit(
-                    channel,
-                    number,
-                    build_14_bit_value_from_two_7_bit_values(value_msb, value_lsb),
-                ),
-                None => ParameterNumberMessage::non_registered_7_bit(channel, number, value_msb),
+        let msg = match self.value_lsb {
+            Some(value_lsb) => ParameterNumberMessage::fourteen_bit(
+                channel,
+                number,
+                build_14_bit_value_from_two_7_bit_values(value_msb, value_lsb),
+                self.is_registered,
+            ),
+            None => {
+                ParameterNumberMessage::seven_bit(channel, number, value_msb, self.is_registered)
             }
         };
         Some(msg)
@@ -186,7 +181,28 @@ mod tests {
     }
 
     #[test]
-    fn should_return_parameter_number_result_message_on_fourth_short_message() {
+    fn x_y_msb() {
+        // Given
+        let mut scanner = ParameterNumberMessageScanner::new();
+        // When
+        let result_1 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(99), u7(3)));
+        let result_2 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(98), u7(37)));
+        let result_3 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(6), u7(126)));
+        // Then
+        assert_eq!(result_1, None);
+        assert_eq!(result_2, None);
+        assert_eq!(
+            result_3,
+            Some(ParameterNumberMessage::non_registered_7_bit(
+                ch(2),
+                u14(421),
+                u7(126)
+            ))
+        );
+    }
+
+    #[test]
+    fn x_y_lsb_msb() {
         // Given
         let mut scanner = ParameterNumberMessageScanner::new();
         // When
@@ -209,13 +225,14 @@ mod tests {
     }
 
     #[test]
-    fn should_return_parameter_number_result_message_on_third_short_message() {
+    fn x_y_msb_msb() {
         // Given
         let mut scanner = ParameterNumberMessageScanner::new();
         // When
         let result_1 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(99), u7(3)));
         let result_2 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(98), u7(37)));
         let result_3 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(6), u7(126)));
+        let result_4 = scanner.feed(&RawShortMessage::control_change(ch(2), cn(6), u7(125)));
         // Then
         assert_eq!(result_1, None);
         assert_eq!(result_2, None);
@@ -225,6 +242,48 @@ mod tests {
                 ch(2),
                 u14(421),
                 u7(126)
+            ))
+        );
+        assert_eq!(
+            result_4,
+            Some(ParameterNumberMessage::non_registered_7_bit(
+                ch(2),
+                u14(421),
+                u7(125)
+            ))
+        );
+    }
+
+    #[test]
+    fn x_y_lsb_msb_lsb_msb() {
+        // Given
+        let mut scanner = ParameterNumberMessageScanner::new();
+        // When
+        let result_1 = scanner.feed(&RawShortMessage::control_change(ch(0), cn(101), u7(3)));
+        let result_2 = scanner.feed(&RawShortMessage::control_change(ch(0), cn(100), u7(36)));
+        let result_3 = scanner.feed(&RawShortMessage::control_change(ch(0), cn(38), u7(24)));
+        let result_4 = scanner.feed(&RawShortMessage::control_change(ch(0), cn(6), u7(117)));
+        let result_5 = scanner.feed(&RawShortMessage::control_change(ch(0), cn(38), u7(23)));
+        let result_6 = scanner.feed(&RawShortMessage::control_change(ch(0), cn(6), u7(117)));
+        // Then
+        assert_eq!(result_1, None);
+        assert_eq!(result_2, None);
+        assert_eq!(result_3, None);
+        assert_eq!(
+            result_4,
+            Some(ParameterNumberMessage::registered_14_bit(
+                ch(0),
+                u14(420),
+                u14(15000)
+            ))
+        );
+        assert_eq!(result_5, None);
+        assert_eq!(
+            result_6,
+            Some(ParameterNumberMessage::registered_14_bit(
+                ch(0),
+                u14(420),
+                u14(14999)
             ))
         );
     }